requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
dnspython==2.4.2
aiohttp==3.9.1
python-multipart==0.0.6